# gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateInventoryErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Create Inventory!
        type: string
    type: object
  models.CreateInventoryPayload:
    properties:
      available_quantity:
        example: 100
        type: integer
      threshold_quantity:
        example: 10
        type: integer
    required:
    - available_quantity
    - threshold_quantity
    type: object
  models.CreateInventorySuccessResponse:
    properties:
      inventory:
        $ref: '#/definitions/models.Inventory'
      message:
        example: Inventory Created Successfully!
        type: string
    type: object
  models.DeleteInventoryByProductIDAndVendorIDErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Delete Inventory!
        type: string
    type: object
  models.DeleteInventoryByProductIDAndVendorIDSuccessResponse:
    properties:
      message:
        example: Inventory Deleted Successfully!
        type: string
    type: object
...
  models.GetInventoryByProductIDAndVendorIDErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Get Inventory!
        type: string
    type: object
  models.GetInventoryByProductIDAndVendorIDSuccessResponse:
    properties:
      inventory:
        $ref: '#/definitions/models.Inventory'
      message:
        example: Inventory Fetched Successfully!
        type: string
    type: object
...
  models.Inventory:
    properties:
      available_quantity:
        example: 100
        type: integer
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      id:
        example: 67adca0b4e864cfffb002299
        type: string
      product_id:
        example: 67adca0b4e864cfffb002299
        type: string
      threshold_quantity:
        example: 10
        type: integer
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      vendor_id:
        example: 67adca0b4e864cfffb002299
        type: string
    type: object
...
  models.UnauthorizedErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Unauthorized!
        type: string
    type: object
  models.UpdateInventoryErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Update Inventory!
        type: string
    type: object
  models.UpdateInventoryPayload:
    properties:
      available_quantity:
        example: 100
        type: integer
      threshold_quantity:
        example: 10
        type: integer
    required:
    - available_quantity
    - threshold_quantity
    type: object
  models.UpdateInventorySuccessResponse:
    properties:
      inventory:
        $ref: '#/definitions/models.Inventory'
      message:
        example: Inventory Updated Successfully!
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: rohit.vilas.ingole@gmail.com
    name: Rohit Vilas Ingole
  description: This is the API for Gopher Social, a social media platform for Gophers.
  license:
    name: MIT License
    url: https://github.com/DataRohit/Order-Management-System/blob/master/license
  title: Gopher Social API
  version: "1.0"
paths:
...
  /product/{id}/inventory:
    delete:
      consumes:
      - application/json
      description: Delete inventory for a product by vendor
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Inventory Deleted Successfully!
          schema:
            $ref: '#/definitions/models.DeleteInventoryByProductIDAndVendorIDSuccessResponse'
        "400":
          description: Failed to Delete Inventory!
          schema:
            $ref: '#/definitions/models.DeleteInventoryByProductIDAndVendorIDErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.DeleteInventoryByProductIDAndVendorIDErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete inventory by product ID
      tags:
      - Inventory
    get:
      consumes:
      - application/json
      description: Get inventory for a product by vendor
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Inventory Fetched Successfully!
          schema:
            $ref: '#/definitions/models.GetInventoryByProductIDAndVendorIDSuccessResponse'
        "400":
          description: Failed to Get Inventory!
          schema:
            $ref: '#/definitions/models.GetInventoryByProductIDAndVendorIDErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.GetInventoryByProductIDAndVendorIDErrorResponse'
      security:
      - BearerAuth: []
      summary: Get inventory by product ID
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      description: Create a new inventory for a product by vendor
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Inventory Create Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.CreateInventoryPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Inventory Created Successfully!
          schema:
            $ref: '#/definitions/models.CreateInventorySuccessResponse'
        "400":
          description: Failed to Create Inventory!
          schema:
            $ref: '#/definitions/models.CreateInventoryErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.CreateInventoryErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new inventory
      tags:
      - Inventory
    put:
      consumes:
      - application/json
      description: Update inventory for a product by vendor
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Inventory Update Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UpdateInventoryPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Inventory Updated Successfully!
          schema:
            $ref: '#/definitions/models.UpdateInventorySuccessResponse'
        "400":
          description: Failed to Update Inventory!
          schema:
            $ref: '#/definitions/models.UpdateInventoryErrorResponse'
        "401":
          description: Unauthorized!
          schema:
            $ref: '#/definitions/models.UnauthorizedErrorResponse'
        "500":
          description: Internal Server Error!
          schema:
            $ref: '#/definitions/models.UpdateInventoryErrorResponse'
      security:
      - BearerAuth: []
      summary: Update inventory
      tags:
      - Inventory
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"